+++ rustls/src/client/tls13.rs
@@ -4,11 +4,11 @@ use crate::conn::Protocol;
 #[cfg(feature = "secret_extraction")]
 use crate::conn::Side;
 use crate::conn::{self, CommonState, ConnectionRandoms, State};
+use crate::crypto::ring::KeyExchange;
 use crate::crypto::CryptoProvider;
 use crate::enums::{ProtocolVersion, SignatureScheme};
 use crate::error::{Error, InvalidMessage, PeerIncompatible, PeerMisbehaved};
 use crate::hash_hs::{HandshakeHash, HandshakeHashBuffer};
-use crate::kx;
 #[cfg(feature = "logging")]
 use crate::log::{debug, trace, warn};
 use crate::msgs::base::{Payload, PayloadU8};
@@ -72,7 +72,7 @@ pub(super) fn handle_server_hello(
     transcript: HandshakeHash,
     early_key_schedule: Option<KeyScheduleEarly>,
     hello: ClientHelloDetails,
-    our_key_share: kx::KeyExchange,
+    our_key_share: KeyExchange,
     mut sent_tls13_fake_ccs: bool,
 ) -> hs::NextStateOrError {
     validate_server_hello(cx.common, server_hello)?;
@@ -187,7 +187,7 @@ fn validate_server_hello(
 pub(super) fn initial_key_share(
     config: &ClientConfig<impl CryptoProvider>,
     server_name: &ServerName,
-) -> Result<kx::KeyExchange, Error> {
+) -> Result<KeyExchange, Error> {
     let group = config
         .session_storage
         .kx_hint(server_name)
@@ -205,7 +205,7 @@ pub(super) fn initial_key_share(
                 .name,
         );
 
-    kx::KeyExchange::choose(group, &config.kx_groups).map_err(|_| Error::FailedToGetRandomBytes)
+    KeyExchange::choose(group, &config.kx_groups).map_err(|_| Error::FailedToGetRandomBytes)
 }
 
 /// This implements the horrifying TLS1.3 hack where PSK binders have a
