+++ rustls/src/lib.rs
@@ -342,7 +342,6 @@ mod enums;
 mod key;
 mod key_log;
 mod key_log_file;
-mod kx;
 mod suites;
 mod ticketer;
 mod versions;
@@ -372,12 +371,12 @@ pub use crate::builder::{
 pub use crate::conn::{
     CommonState, Connection, ConnectionCommon, IoState, Reader, Side, SideData, Writer,
 };
+pub use crate::crypto::ring::{SupportedKxGroup, ALL_KX_GROUPS};
 pub use crate::enums::{CipherSuite, ProtocolVersion, SignatureScheme};
 pub use crate::error::{CertificateError, Error, InvalidMessage, PeerIncompatible, PeerMisbehaved};
 pub use crate::key::{Certificate, PrivateKey};
 pub use crate::key_log::{KeyLog, NoKeyLog};
 pub use crate::key_log_file::KeyLogFile;
-pub use crate::kx::{SupportedKxGroup, ALL_KX_GROUPS};
 pub use crate::msgs::enums::{
     AlertDescription, ContentType, HandshakeType, NamedGroup, SignatureAlgorithm,
 };
@@ -496,9 +495,9 @@ pub mod version {
 ///
 /// ALL_KX_GROUPS is provided as an array of all of these values.
 pub mod kx_group {
-    pub use crate::kx::SECP256R1;
-    pub use crate::kx::SECP384R1;
-    pub use crate::kx::X25519;
+    pub use crate::crypto::ring::SECP256R1;
+    pub use crate::crypto::ring::SECP384R1;
+    pub use crate::crypto::ring::X25519;
 }
 
 /// Message signing interfaces and implementations.
