+++ rustls/src/tls12/mod.rs
@@ -1,8 +1,8 @@
 use crate::cipher::{MessageDecrypter, MessageEncrypter};
 use crate::conn::{CommonState, ConnectionRandoms, Side};
+use crate::crypto;
 use crate::enums::{CipherSuite, SignatureScheme};
 use crate::error::{Error, InvalidMessage};
-use crate::kx;
 use crate::msgs::codec::{Codec, Reader};
 use crate::msgs::enums::AlertDescription;
 use crate::msgs::handshake::KeyExchangeAlgorithm;
@@ -206,7 +206,7 @@ pub(crate) struct ConnectionSecrets {
 
 impl ConnectionSecrets {
     pub(crate) fn from_key_exchange(
-        kx: kx::KeyExchange,
+        kx: crypto::ring::KeyExchange,
         peer_pub_key: &[u8],
         ems_seed: Option<Digest>,
         randoms: ConnectionRandoms,
@@ -515,11 +515,12 @@ pub(crate) const DOWNGRADE_SENTINEL: [u8; 8] = [0x44, 0x4f, 0x57, 0x4e, 0x47, 0x
 mod tests {
     use super::*;
     use crate::conn::{CommonState, Side};
+    use crate::crypto::ring::{KeyExchange, X25519};
     use crate::msgs::handshake::{ClientECDHParams, ServerECDHParams};
 
     #[test]
     fn server_ecdhe_remaining_bytes() {
-        let key = kx::KeyExchange::start(&kx::X25519).unwrap();
+        let key = KeyExchange::start(&X25519).unwrap();
         let server_params = ServerECDHParams::new(key.group(), key.pub_key());
         let mut server_buf = Vec::new();
         server_params.encode(&mut server_buf);
